// src/app/interfaces/user.ts
import { Address } from '../classes/address';

export interface User {
    id: string;
    firstname: string;
    lastname: string;
    email: string;
    idNumber: string;
    age: string;
    qualification: string;
    employmentStatus: string;
    votingStation: string;
    photoUrl: string;
    quote: string;
    bio: string;
}